+++ src/dodal/devices/electron_analyser/types.py
@@ -1,6 +1,23 @@
-from ophyd_async.core import StrictEnum
+from dodal.devices.electron_analyser.abstract.base_driver_io import (
+    AbstractAnalyserDriverIO,
+)
+from dodal.devices.electron_analyser.abstract.base_region import (
+    AbstractBaseRegion,
+    AbstractBaseSequence,
+)
+from dodal.devices.electron_analyser.detector import (
+    ElectronAnalyserDetector,
+    ElectronAnalyserRegionDetector,
+)
+from dodal.devices.electron_analyser.specs.detector import SpecsDetector
+from dodal.devices.electron_analyser.vgscienta.detector import VGScientaDetector
 
+ElectronAnalyserDetectorImpl = VGScientaDetector | SpecsDetector
 
-class EnergyMode(StrictEnum):
-    KINETIC = "Kinetic"
-    BINDING = "Binding"
+GenericElectronAnalyserDetector = ElectronAnalyserDetector[
+    AbstractAnalyserDriverIO, AbstractBaseSequence, AbstractBaseRegion
+]
+
+GenericElectronAnalyserRegionDetector = ElectronAnalyserRegionDetector[
+    AbstractAnalyserDriverIO, AbstractBaseRegion
+]
